Tourism Regions and Nearest Weather Stations

A dataset containing the locations of Australian tourism regions, their geographic coordinates, and the nearest Bureau of Meteorology weather station. Each region is assigned a unique identifier for linking to other tourism datasets.

Usage

tourism_region

Format

A data frame with `r nrow(tourism_region)` rows and 5 variables: * **region**: Name of the tourism region. Tourism regions are defined by Tourism Research Australia and generally formed through the aggregation of Statistical Local Areas (SLAs) or other ABS-defined geographies. * **lon**: Longitude of the tourism region representative point (WGS84). * **lat**: Latitude of the tourism region representative point (WGS84). * **region_id**: Unique integer identifier for the tourism region. Useful for joining with other tourism-related datasets. * **ws_id**: Identifier of the nearest Bureau of Meteorology weather station to the tourism region.

Details

Coordinates for each tourism region are intended to represent a central location within the region (e.g., polygon centroid). The nearest weather station is determined using great-circle distance calculations based on the Bureau of Meteorology's official station list.

References

Tourism Research Australia: https://www.tra.gov.au Australian Bureau of Meteorology: http://www.bom.gov.au

Examples

data(tourism_region)
head(tourism_region)
#> # A tibble: 6 × 5
#>   region             lon   lat region_id ws_id       
#>   <chr>            <dbl> <dbl>     <int> <chr>       
#> 1 ACT - South West  149. -35.4         1 949250-99999
#> 2 APY Lands         131. -27.1         2 944740-99999
#> 3 Abbotsford        145. -37.8         3 959360-99999
#> 4 Aberfoyle Park    139. -35.1         4 946830-99999
#> 5 Acacia Gardens    151. -33.7         5 947640-99999
#> 6 Acton             145. -37.9         6 948650-99999